--- a/src/test/java/io/tarantool/driver/benchmark/ClusterBenchmarkRunner.java
+++ b/src/test/java/io/tarantool/driver/benchmark/ClusterBenchmarkRunner.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
@@ -23,35 +24,61 @@ import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.TearDown;
+import org.openjdk.jmh.annotations.Warmup;
 import org.openjdk.jmh.infra.Blackhole;
 
-@Fork(value = 1, jvmArgsAppend = "-Xmx1G")
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+@Fork(value = 1, jvmArgsAppend = {
+    "-Xms1G", "-Xmx1G", "-XX:+UseG1GC", //"-XX:+UnlockCommercialFeatures",
+    "-XX:+FlightRecorder",
+    "--add-opens=java.base/java.nio=ALL-UNNAMED", "--add-opens=java.base/sun.nio.ch=ALL-UNNAMED",
+    "-XX:StartFlightRecording=settings=/home/fedora/sources/cartridge-java/profile.jfc"
+})
 public class ClusterBenchmarkRunner {
     private static final String TEST_SPACE = "test_space";
     private static final String TEST_PROFILE = "test__profile";
+    private static final String CLUSTER_INIT_SCRIPT = "org/testcontainers/containers/cluster_benchmark.lua";
+
+    private static Logger logger = LoggerFactory.getLogger(ClusterBenchmarkRunner.class);
 
     public static void main(String[] args) throws Exception {
         org.openjdk.jmh.Main.main(args);
     }
 
     @Benchmark
+    @BenchmarkMode(Mode.Throughput)
+    @Warmup(iterations = 2)
     @Measurement(iterations = 10)
-    @OperationsPerInvocation(2)
-    public void getSpaceObject(ClusterTarantoolSetup plan, Blackhole bh) {
-        TarantoolSpaceOperations<TarantoolTuple, TarantoolResult<TarantoolTuple>> testSpace =
-            plan.tarantoolClient.space(TEST_SPACE);
-        bh.consume(testSpace);
-
-        TarantoolSpaceOperations<TarantoolTuple, TarantoolResult<TarantoolTuple>> profileSpace =
-            plan.tarantoolClient.space(TEST_PROFILE);
-        bh.consume(profileSpace);
+    @OperationsPerInvocation(2000)
+    public void writeDataUsingCallAPI(ClusterTarantoolSetup plan, FuturesHolder futuresHolder) {
+        // Fill 10000 rows into both spaces
+        TarantoolTuple tarantoolTuple;
+        String uuid;
+        int nextId = 0;
+        for (int i = 0; i < 1_000; i++) {
+            uuid = UUID.randomUUID().toString();
+            nextId = plan.nextTestSpaceId + i;
+            tarantoolTuple = plan.tupleFactory.create(1_000_000 + nextId, null, uuid, 200_000 + nextId);
+            futuresHolder.allFutures.add(plan.tarantoolClient.callForSingleResult(
+                "custom_crud_insert_one_record" , Arrays.asList(TEST_SPACE, tarantoolTuple), Map.class));
+            tarantoolTuple = plan.tupleFactory.create(1_000_000 + nextId, null, uuid, 50_000 + nextId, 100_000 + i);
+            futuresHolder.allFutures.add(plan.tarantoolClient.callForSingleResult(
+                "custom_crud_insert_one_record" , Arrays.asList(TEST_PROFILE, tarantoolTuple), Map.class));
+        }
+        nextId++;
+        plan.nextTestSpaceId = nextId;
+        plan.nextProfileSpaceId = nextId;
     }
 
     @Benchmark
     @BenchmarkMode(Mode.Throughput)
+    @Warmup(iterations = 2)
     @Measurement(iterations = 10)
     @OperationsPerInvocation(2000)
-    public void writeData(ClusterTarantoolSetup plan, FuturesHolder futuresHolder, Spaces spaces, Blackhole bh) {
+    public void writeDataUsingSpaceAPI(
+        ClusterTarantoolSetup plan, FuturesHolder futuresHolder, Spaces spaces) {
         // Fill 10000 rows into both spaces
         TarantoolTuple tarantoolTuple;
         String uuid;
@@ -71,14 +98,18 @@ public class ClusterBenchmarkRunner {
 
     @Benchmark
     @BenchmarkMode(Mode.Throughput)
+    @Warmup(iterations = 2)
     @Measurement(iterations = 10)
     @OperationsPerInvocation(1000)
-    public void readDataUsingCallAPI(ClusterTarantoolSetup plan, FuturesHolder futuresHolder, Blackhole bh) {
-        boolean coin = Math.random() - 0.5 > 0;
-        String spaceName = coin ? TEST_SPACE : TEST_PROFILE;
+    public void readDataUsingCallAPI(
+        ClusterTarantoolSetup plan, FuturesHolder futuresHolder, ClusterDataInit clusterDataInit) {
+        boolean coin;
+        String spaceName;
         long nextId;
         for (int i = 0; i < 1_000; i++) {
             nextId = Math.round(Math.random() * 10_000) + 1_000_000;
+            coin = Math.random() - 0.5 > 0;
+            spaceName = coin ? TEST_SPACE : TEST_PROFILE;
             futuresHolder.allFutures.add(
                 plan.tarantoolClient.callForSingleResult(
                     "custom_crud_get_one_record", Arrays.asList(spaceName, nextId), List.class)
@@ -88,17 +119,20 @@ public class ClusterBenchmarkRunner {
 
     @Benchmark
     @BenchmarkMode(Mode.Throughput)
+    @Warmup(iterations = 2)
     @Measurement(iterations = 10)
     @OperationsPerInvocation(1000)
     public void readDataUsingSpaceAPI(
-        ClusterTarantoolSetup plan, FuturesHolder futuresHolder, Spaces spaces, Blackhole bh) {
-        boolean coin = Math.random() - 0.5 > 0;
-        TarantoolSpaceOperations<TarantoolTuple, TarantoolResult<TarantoolTuple>> space = coin ?
-            spaces.testSpace : spaces.profileSpace;
-        String pkFieldName = coin ? "id" : "profile_id";
+        ClusterTarantoolSetup plan, FuturesHolder futuresHolder, Spaces spaces, ClusterDataInit clusterDataInit) {
+        boolean coin;
+        TarantoolSpaceOperations<TarantoolTuple, TarantoolResult<TarantoolTuple>> space;
+        String pkFieldName;
         long nextId;
         for (int i = 0; i < 1_000; i++) {
             nextId = Math.round(Math.random() * 10_000) + 1_000_000;
+            coin = Math.random() - 0.5 > 0;
+            space = coin ? spaces.testSpace : spaces.profileSpace;
+            pkFieldName = coin ? "id" : "profile_id";
             futuresHolder.allFutures.add(
                 space.select(Conditions.indexEquals(pkFieldName, Collections.singletonList(nextId)))
             );
@@ -109,14 +143,14 @@ public class ClusterBenchmarkRunner {
     public static class FuturesHolder {
         final List<CompletableFuture<?>> allFutures = new ArrayList<>(2_000);
 
-        @Setup(Level.Invocation)
-        public void doSetup() {
-            allFutures.clear();
-        }
-
         @TearDown(Level.Invocation)
         public void doTeardown() {
-            allFutures.forEach(CompletableFuture::join);
+            try {
+                allFutures.forEach(CompletableFuture::join);
+                allFutures.clear();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
     }
 
@@ -129,6 +163,15 @@ public class ClusterBenchmarkRunner {
         public void doSetup(ClusterTarantoolSetup plan) {
             testSpace = plan.tarantoolClient.space(TEST_SPACE);
             profileSpace = plan.tarantoolClient.space(TEST_PROFILE);
+        }
+    }
+
+    @State(Scope.Thread)
+    public static class ClusterDataInit {
+
+        @Setup(Level.Iteration)
+        public void doSetup(ClusterTarantoolSetup plan) throws Exception {
+            plan.tarantoolContainer.executeScript(CLUSTER_INIT_SCRIPT);
         }
     }
 }