--- a/src/test/resources/cartridge/app/roles/api_router.lua
+++ b/src/test/resources/cartridge/app/roles/api_router.lua
@@ -100,6 +100,10 @@ local function custom_crud_get_one_record(space_name, tuple_id)
     return crud.get(space_name, {tuple_id}).rows[1]
 end
 
+local function custom_crud_insert_one_record(space_name, tuple)
+    return crud.insert(space_name, tuple)
+end
+
 local function raising_error()
     error("Test error: raising_error() called")
 end
@@ -251,6 +255,7 @@ local function init(opts)
     rawset(_G, 'crud_error_timeout', crud_error_timeout)
     rawset(_G, 'custom_crud_select', custom_crud_select)
     rawset(_G, 'custom_crud_get_one_record', custom_crud_get_one_record)
+    rawset(_G, 'custom_crud_insert_one_record', custom_crud_insert_one_record)
     rawset(_G, 'get_routers_status', get_routers_status)
     rawset(_G, 'init_router_status', init_router_status)
     rawset(_G, 'test_no_such_procedure', test_no_such_procedure)